+++ CHANGELOG.md
@@ -35,6 +35,7 @@ migrate tests from prior versions.
 - **[BC]** The function passed to `Call()` no longer returns an `error`
 - **[BC]** `engine.New()` and `MustNew()` now accept `configkit.RichApplication` (previously `dogma.Application`)
 - **[BC]** Rename `WithStartTime()` to `StartVirtualClockAt()`
+- **[BC]** Rename `WithOperationOptions()` to `WithUnsafeOperationOptions()`
 
 ### Removed